pyproject.toml
[project]
name = "gpalytics-backend"
version = "0.1.0"
description = "Modern FastAPI backend for GPA analytics with Azure SQL and containerization"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.14",
    "sqlalchemy[asyncio]>=2.0.0",
    "aioodbc>=0.4.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "PyJWT>=2.8.0",
    "argon2-cffi>=23.1.0",
    "pytest>=8.4.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.10.1",
    "aiosqlite>=0.21.0",
    "pytesseract>=0.3.13",
    "pillow>=11.3.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
    "pyodbc>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
